pyproject.toml
[project]
name = "langchain-pangea"
version = "0.0.0"
description = "LangChain integrations for Pangea"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Pangea", email = "support@pangea.cloud" }
]
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.61",
    "pangea-sdk>=6.1.1",
    "pydantic>=2.11.5",
    "typing-extensions>=4.13.2",
]

[project.urls]
"Source Code" = "https://github.com/pangeacyber/langchain-pangea"
"Release Notes" = "https://github.com/pangeacyber/langchain-pangea/releases"
repository = "https://github.com/pangeacyber/langchain-pangea"

[dependency-groups]
dev = [
    "langchain-tests==0.3.19",
    "mypy==1.15.0",
    "pytest==8.3.5",
    "ruff==0.11.11",
]

[build-system]
requires = ["pdm-backend==2.4.4"]
build-backend = "pdm.backend"

[tool.mypy]
disallow_untyped_defs = true

[tool.pytest.ini_options]
addopts = "--tb=short"
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
filterwarnings = ["error"]
testpaths = ["tests"]
xfail_strict = true

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",   # isort
    "PTH", # flake8-use-pathlib
    "TC",  # flake8-type-checking
    "UP",  # pyupgrade
]
flake8-type-checking.runtime-evaluated-base-classes = [
    "pydantic.BaseModel",
    "langchain_core.load.serializable.Serializable",
    "langchain_core.runnables.base.RunnableSerializable"
]
